gkeygen/actions.py: Message updates to show spec file and url being used
Change a logger.error message to a logger.debug as the error info printed was too generic to be useful.
Tweak the input messages for better display.
etc/gkeys-gen.conf: Add spec-homepage section

In order to make the creator fully configurable.
We needed to make the homepage for the spec configurable too.