gkeys/fetch.py: New Fetch class
Add sig_path assignment, makes the sig_path available to class instance holders like _verify()
Use the lack of a returned timestamp to abort the process
If the main file is not downloaded, then don't bother looking for a signature.
This allows a verify action to exit smoothly.
There is a message about the reason for not downloading, so is not hidden.

Signed-off-by: Brian Dolbec <dolsen@gentoo.org>